# weyyak_gcp/staging/api/content/docs/swagger.yaml
definitions:
...
  productname.ProductName:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/content/productnames:
    get:
      consumes:
      - application/json
      description: get product names
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/productname.ProductName'
            type: array
      security:
      - Authorization: []
      summary: Show ProductNames
      tags:
      - Product Name
...
  /api/contentvariances/{id}:
    delete:
      consumes:
      - application/json
      description: Delete content variances details by contentvariance id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Delete content variances details by contentvariance id
      tags:
      - contentvariance
  /api/contentvariances/{id}/status/{status}:
    post:
      consumes:
      - application/json
      description: post Update  contentvariance status
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      - description: status
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Show contentvariance status
      tags:
      - contentvariance
...
  /api/episodes:
    get:
      consumes:
      - application/json
      description: Get Episodes Based on Season ID
      parameters:
      - description: seasonid
        in: path
        name: seasonid
        type: string
      - description: Offset
        in: query
        name: offset
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: Page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Get Episodes Based on Season ID
      tags:
      - Episode
  /api/episodes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete episode details by episode id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Delete episode details by episode id
      tags:
      - episode
  /api/episodes/{id}/status/{status}:
    post:
      consumes:
      - application/json
      description: post Update Episode status
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      - description: status
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Show Update Episode status
      tags:
      - episode
...
  /api/seasons:
    get:
      consumes:
      - application/json
      description: Get List of Seasons from content id
      parameters:
      - description: Content Id
        in: path
        name: ContentId
        type: string
      - description: Offset
        in: query
        name: offset
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: Page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Get List of Seasons from content id
      tags:
      - Season
  /api/seasons/{id}/status/{status}:
    post:
      consumes:
      - application/json
      description: post Update Season
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      - description: status
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - Authorization: []
      summary: Show Update Season
      tags:
      - Season
...
swagger: "2.0"